refactor(sidebar): icon-only filter button with active-dot indicator
Split the filter back out of view options into a standalone icon
button. Drop the text chip — the dropdown itself shows which repos /
filters are applied when opened. An active-state dot on the icon
preserves at-a-glance visibility that filtering is on.
